Moved encoding/decoding test function to xml_utils.cpp files and deleted unused libraries
