+++ src/rcpp_parse_chromatograms.cpp
@@ -2,18 +2,7 @@
 #include "external_libraries.hpp"
 #include "xml_utils.h"
 #include <string>
-#include <vector>
-#include <list>
 #include <Rcpp.h>
-#include <cstdio>
-#include <cstring>
-#include <iomanip>
-#include <sstream>
-#include <stdexcept>
-#include <zlib.h>
-#include <algorithm>
-#include <iterator>
-
 
 
 // [[Rcpp::export]]
